+++ src/glslang.rs
@@ -149,8 +149,8 @@ impl Glslang {
                         _ => ShaderErrorSeverity::Error,
                     },
                     error: String::from(msg),
-                    line: line.parse::<usize>().unwrap_or(1),
-                    pos: pos.parse::<usize>().unwrap_or(0),
+                    line: line.parse::<u32>().unwrap_or(1),
+                    pos: pos.parse::<u32>().unwrap_or(0),
                 });
             } else {
                 shader_error_list.push(ShaderError::InternalErr(format!(
